feat(menu): browser zoom (#625)
* feat(menu): rename file

* feat(menu): add browser zoom

* feat(menu): use helper methods

* feat(menu): break menu into individual functions

* feat(menu): fixed windows

* feat(zoom): added mouse wheel event handler to preload renderer
